Output
This python program will show the highest palindrome number that is a product of two numbers of n digits. 
Piece of advice: if you are looking for the highest palindrome that is a product of 3 digit numbers, set the lowest limit 
to 9900 and highest limit to 10000.
In order to speed the process, work with the lowest possible range.
Note. A palindromic number reads the same both ways.
Lowest limit of the range: 900
Highest limit of the range: 1000
888888 is the product of 924 times 962.

This python program will show the highest palindrome number that is a product of two numbers of n digits. 
Piece of advice: if you are looking for the highest palindrome that is a product of 3 digit numbers, set the lowest limit 
to 9900 and highest limit to 10000.
In order to speed the process, work with the lowest possible range.
Note. A palindromic number reads the same both ways.
Lowest limit of the range: 99900
Highest limit of the range: 100000
There are no such numbers in this range. Try a lowest limit, maybe 98900.

This python program will show the highest palindrome number that is a product of two numbers of n digits. 
Piece of advice: if you are looking for the highest palindrome that is a product of 3 digit numbers, set the lowest limit 
to 9900 and highest limit to 10000.
In order to speed the process, work with the lowest possible range.
Note. A palindromic number reads the same both ways.
Lowest limit of the range: 98900
Highest limit of the range: 100000
9966006699 is the product of 99681 times 99979.